Move reconstruction IO to a new file
