+++ src/colmap/scene/reconstruction.h
@@ -129,6 +129,9 @@ class Reconstruction {
   // Add new image.
   void AddImage(class Image image);
 
+  // Add new 3D point with known ID.
+  void AddPoint3D(point3D_t point3D_id, struct Point3D point3D);
+
   // Add new 3D object, and return its unique ID.
   point3D_t AddPoint3D(
       const Eigen::Vector3d& xyz,
@@ -260,85 +263,9 @@ class Reconstruction {
   std::vector<PlyPoint> ConvertToPLY() const;
 
   // Import from other data formats. Note that these import functions are
-  // only intended for visualization of data and usable for reconstruction.
+  // only intended for visualization of data and unusable for reconstruction.
   void ImportPLY(const std::string& path);
   void ImportPLY(const std::vector<PlyPoint>& ply_points);
-
-  // Export to other data formats.
-
-  // Exports in NVM format http://ccwu.me/vsfm/doc.html#nvm. Only supports
-  // SIMPLE_RADIAL camera model when exporting distortion parameters. When
-  // skip_distortion == true it supports all camera models with the caveat that
-  // it's using the mean focal length which will be inaccurate for camera models
-  // with two focal lengths and distortion.
-  bool ExportNVM(const std::string& path, bool skip_distortion = false) const;
-
-  // Exports in CAM format which is a simple text file that contains pose
-  // information and camera intrinsics for each image and exports one file per
-  // image; it does not include information on the 3D points. The format is as
-  // follows (2 lines of text with space separated numbers):
-  // <Tvec; 3 values> <Rotation matrix in row-major format; 9 values>
-  // <focal_length> <k1> <k2> 1.0 <principal point X> <principal point Y>
-  // Note that focal length is relative to the image max(width, height),
-  // and principal points x and y are relative to width and height respectively.
-  //
-  // Only supports SIMPLE_RADIAL and RADIAL camera models when exporting
-  // distortion parameters. When skip_distortion == true it supports all camera
-  // models with the caveat that it's using the mean focal length which will be
-  // inaccurate for camera models with two focal lengths and distortion.
-  bool ExportCam(const std::string& path, bool skip_distortion = false) const;
-
-  // Exports in Recon3D format which consists of three text files with the
-  // following format and content:
-  // 1) imagemap_0.txt: a list of image numeric IDs with one entry per line.
-  // 2) urd-images.txt: A list of images with one entry per line as:
-  //    <image file name> <width> <height>
-  // 3) synth_0.out: Contains information for image poses, camera intrinsics,
-  //    and 3D points as:
-  //    <N; num images> <M; num points>
-  //    <N lines of image entries>
-  //    <M lines of point entries>
-  //
-  //    Each image entry consists of 5 lines as:
-  //    <focal length> <k1> <k2>
-  //    <Rotation matrix; 3x3 array>
-  //    <Tvec; 3 values>
-  //    Note that the focal length is scaled by 1 / max(width, height)
-  //
-  //    Each point entry consists of 3 lines as:
-  //    <point x, y, z coordinates>
-  //    <point RGB color>
-  //    <K; num track elements> <Track Element 1> ... <Track Element K>
-  //
-  //    Each track elemenet is a sequence of 5 values as:
-  //    <image ID> <2D point ID> -1.0 <X> <Y>
-  //    Note that the 2D point coordinates are centered around the principal
-  //    point and scaled by 1 / max(width, height).
-  //
-  // When skip_distortion == true it supports all camera models with the
-  // caveat that it's using the mean focal length which will be inaccurate
-  // for camera models with two focal lengths and distortion.
-  bool ExportRecon3D(const std::string& path,
-                     bool skip_distortion = false) const;
-
-  // Exports in Bundler format https://www.cs.cornell.edu/~snavely/bundler/.
-  // Supports SIMPLE_PINHOLE, PINHOLE, SIMPLE_RADIAL and RADIAL camera models
-  // when exporting distortion parameters. When skip_distortion == true it
-  // supports all camera models with the caveat that it's using the mean focal
-  // length which will be inaccurate for camera models with two focal lengths
-  // and distortion.
-  bool ExportBundler(const std::string& path,
-                     const std::string& list_path,
-                     bool skip_distortion = false) const;
-
-  // Exports 3D points only in PLY format.
-  void ExportPLY(const std::string& path) const;
-
-  // Exports in VRML format https://en.wikipedia.org/wiki/VRML.
-  void ExportVRML(const std::string& images_path,
-                  const std::string& points3D_path,
-                  double image_scale,
-                  const Eigen::Vector3d& image_rgb) const;
 
   // Extract colors for 3D points of given image. Colors will be extracted
   // only for 3D points which are completely black.
@@ -371,20 +298,6 @@ class Reconstruction {
   std::tuple<Eigen::Vector3d, Eigen::Vector3d, Eigen::Vector3d>
   ComputeBoundsAndCentroid(double p0, double p1, bool use_images) const;
 
-  void ReadCamerasText(const std::string& path);
-  void ReadImagesText(const std::string& path);
-  void ReadPoints3DText(const std::string& path);
-  void ReadCamerasBinary(const std::string& path);
-  void ReadImagesBinary(const std::string& path);
-  void ReadPoints3DBinary(const std::string& path);
-
-  void WriteCamerasText(const std::string& path) const;
-  void WriteImagesText(const std::string& path) const;
-  void WritePoints3DText(const std::string& path) const;
-  void WriteCamerasBinary(const std::string& path) const;
-  void WriteImagesBinary(const std::string& path) const;
-  void WritePoints3DBinary(const std::string& path) const;
-
   void SetObservationAsTriangulated(image_t image_id,
                                     point2D_t point2D_idx,
                                     bool is_continued_point3D);
@@ -404,7 +317,7 @@ class Reconstruction {
   std::vector<image_t> reg_image_ids_;
 
   // Total number of added 3D points, used to generate unique identifiers.
-  point3D_t num_added_points3D_;
+  point3D_t max_point3D_id_;
 };
 
 ////////////////////////////////////////////////////////////////////////////////
